requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18      
aiosignal==1.3.2      
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1    
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
colorama==0.4.6
comtypes==1.4.10
customtkinter==5.2.2
darkdetect==0.8.0
distro==1.9.0
docopt==0.6.2
frozenlist==1.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.9.0
Js2Py==0.74
multidict==6.4.3
openai==0.28.0
packaging==25.0
pipwin==0.5.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydantic==2.11.3
pydantic_core==2.33.1
pyjsparser==2.7.1
pyparsing==3.2.3
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.1.0
pyttsx3==2.98
pywin32==310
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.2
standard-aifc==3.13.0
standard-chunk==3.13.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
wikipedia==1.4.0
yarl==1.20.0


#pip install -r requirements.txt
#pipwin install pyaudio